pyproject.toml
[project]
name = "dataday-automatic-check-in"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "flask>=3.1.1",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
]
